move AdaptiveParameter from operator to its own module
